+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/routes/index.tsx
@@ -12,4 +12,4 @@
  * <http://www.gnu.org/licenses/lgpl.html>.
  *
  **/
-export { default } from './routes-container'
+export { default } from './routes'
